saw a commit in the wild with more than 65535 additions in a single file zig/b77e6013422a6e7066a14dbe82e8512636ba13d2

+++ git.zig
@@ -635,8 +635,8 @@ pub const TreeDiff = struct {
         after: State,
         action: Action,
         sub_path: string,
-        adds: u16,
-        subs: u16,
+        adds: u32,
+        subs: u32,
     };
 
     pub const State = struct {
